wonderful updates here stolen from lstoll

+++ Rakefile
@@ -4,9 +4,9 @@ desc "Hook our dotfiles into system-standard positions."
 task :install do
   linkables = Dir.glob('**/*.symlink')
 
-  skip_all = false
+  skip_all = !!ENV['SKIP_ALL'] || false
+  backup_all = !!ENV['BACKUP_ALL'] || false
   overwrite_all = false
-  backup_all = false
 
   linkables.each do |linkable|
     overwrite = false
@@ -15,22 +15,27 @@ task :install do
     file = linkable.split('/').last.split('.symlink').last
     target = "#{ENV["HOME"]}/.#{file}"
 
-    if File.exists?(target) || File.symlink?(target)
-      unless skip_all || overwrite_all || backup_all
-        puts "File already exists: #{target}, what do you want to do? [s]kip, [S]kip all, [o]verwrite, [O]verwrite all, [b]ackup, [B]ackup all"
-        case STDIN.gets.chomp
-        when 'o' then overwrite = true
-        when 'b' then backup = true
-        when 'O' then overwrite_all = true
-        when 'B' then backup_all = true
-        when 'S' then skip_all = true
-        when 's' then next
+    if File.exists?(target) && File.symlink?(target) && ENV['PWD'] + '/' + linkable == File.readlink(target)
+      puts "Skipping #{linkable} - already set correctly"
+    else
+      if File.exists?(target) || File.symlink?(target)
+        unless skip_all || overwrite_all || backup_all
+          puts "File already exists: #{target}, what do you want to do? [s]kip, [S]kip all, [o]verwrite, [O]verwrite all, [b]ackup, [B]ackup all"
+          case STDIN.gets.chomp
+          when 'o' then overwrite = true
+          when 'b' then backup = true
+          when 'O' then overwrite_all = true
+          when 'B' then backup_all = true
+          when 'S' then skip_all = true
+          when 's' then next
+          end
         end
+        FileUtils.rm_rf(target) if overwrite || overwrite_all
+        `mv "$HOME/.#{file}" "$HOME/.#{file}.backup"` if backup || backup_all
       end
-      FileUtils.rm_rf(target) if overwrite || overwrite_all
-      `mv "$HOME/.#{file}" "$HOME/.#{file}.backup"` if backup || backup_all
+      puts "Linking #{linkable}"
+      `ln -s "$PWD/#{linkable}" "#{target}"`
     end
-    `ln -s "$PWD/#{linkable}" "#{target}"`
   end
 end
